Refactored pane stress quad unit test to utilize abstract static problem base class
